Make prep script for test files
Previously, used some magic to find the test files in the same directory
as CAMPAREE was installed into. This is hard for the user to see what
is happening and doesn't match what a user will do in practice.
In doing this, we've removed the need for default directories
for resources and fastq files.